fastapi/requirements.txt
fastapi
uvicorn
tensorflow
pydantic
python-jose
python-multipart
Pillow
numpy
sqlalchemy
psycopg2-binary
nltk